Rename hooks to interceptors.
