+++ src/main/java/com/refinedmods/refinedstorage/api/autocrafting/task/interceptor/IOutputInterceptor.java
@@ -1,9 +1,9 @@
-package com.refinedmods.refinedstorage.api.autocrafting.task;
+package com.refinedmods.refinedstorage.api.autocrafting.task.interceptor;
 
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.fluids.FluidStack;
 
-public interface ICraftingTaskOutputHook {
+public interface IOutputInterceptor {
     ItemStack intercept(ItemStack stack);
 
     FluidStack intercept(FluidStack stack);
